requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
celery==5.3.6
stripe==7.6.0
python-decouple==3.8
django-crispy-forms==2.1
crispy-bootstrap5==0.7
Pillow==10.1.0
django-environ==0.11.2
gunicorn==21.2.0 